Refactor: Implement cookie management and disclaimer modal functionality; enhance footer interaction and accessibility
