+++ index.html
@@ -69,7 +69,7 @@
         {
           "@type": "Person",
           "name": "Hassan Saei",
-          "url": "https://github.com/hassansaei",
+          "url": "https://github.com/hassansaei"
         }
       ]
     }
